backend/python server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
livekit==0.12.0
pydantic==2.5.3
google-generativeai==0.8.3
wsproto>=1.2.0
SpeechRecognition==3.10.4

